+++ ESFramework/UIKit/UIControl+ESAdditions.h
@@ -1,0 +1,21 @@
+//
+//  UIControl+ESAdditions.h
+//  ESFramework
+//
+//  Created by Elf Sundae on 2019/05/24.
+//  Copyright © 2019 www.0x123.com. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface UIControl (ESAdditions)
+
+- (void)addActionBlock:(void (^)(__kindof UIControl *control))actionBlock forControlEvents:(UIControlEvents)controlEvents;
+- (void)setActionBlock:(void (^ _Nullable)(__kindof UIControl *control))actionBlock forControlEvents:(UIControlEvents)controlEvents;
+- (void)removeAllActionBlocksForControlEvents:(UIControlEvents)controlEvents;
+
+@end
+
+NS_ASSUME_NONNULL_END