+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `ESActionBlockContainer` class.
 - Added UIBarButtonItem `actionBlock` property.
 - Added UIGestureRecognizer action block methods: `-initWithActionBlock:`, `-addActionBlock:`, `removeAllActionBlocks`.
+- Added UIControl action block methods.
 
 ## 3.1.0 (2019-05-23)